refactor upgrade example
