+++ examples/upgradeable/ignition/modules/ProxyModule.js
@@ -1,4 +1,3 @@
-// ./ignition/LockModule.js
 const { buildModule } = require("@nomicfoundation/hardhat-ignition/modules");
 
 /**
@@ -23,7 +22,7 @@ const proxyModule = buildModule("ProxyModule", (m) => {
     "0x",
   ]);
 
-  // We need to get the address of the ProxyAdmin contrac that was created by the TransparentUpgradeableProxy
+  // We need to get the address of the ProxyAdmin contract that was created by the TransparentUpgradeableProxy
   // so that we can use it to upgrade the proxy later.
   const proxyAdminAddress = m.readEventArgument(
     proxy,
@@ -31,7 +30,7 @@ const proxyModule = buildModule("ProxyModule", (m) => {
     "newAdmin"
   );
 
-  // Here we use m.contractAt(...) to create a contract instance for the ProxyAdmin that we can interact with.
+  // Here we use m.contractAt(...) to create a contract instance for the ProxyAdmin that we can interact with later to upgrade the proxy.
   const proxyAdmin = m.contractAt("ProxyAdmin", proxyAdminAddress);
 
   // Return the proxy and proxy admin so that they can be used by other modules.
@@ -39,47 +38,22 @@ const proxyModule = buildModule("ProxyModule", (m) => {
 });
 
 /**
- * This is the second module that will be run. It upgrades the proxy to a new
- * version of the Demo contract.
+ * This is the second module that will be run, and it is also the only module exported from this file.
+ * It creates a contract instance for the Demo contract using the proxy from the previous module.
  */
-const upgradeModule = buildModule("UpgradeModule", (m) => {
-  // Make sure we're account that owns the ProxyAdmin contract.
-  const proxyAdminOwner = m.getAccount(0);
-
+const demoModule = buildModule("DemoModule", (m) => {
   // Get the proxy and proxy admin from the previous module.
-  const { proxyAdmin, proxy } = m.useModule(proxyModule);
+  const { proxy, proxyAdmin } = m.useModule(proxyModule);
 
-  // This is the new version of the Demo contract that we want to upgrade to.
-  const demoV2 = m.contract("DemoV2");
+  // Here we're using m.contractAt(...) a bit differently than we did above.
+  // While we're still using it to create a contract instance, we're now telling Hardhat Ignition
+  // to treat the contract at the proxy address as an instance of the Demo contract.
+  // This allows us to interact with the underlying Demo contract via the proxy from within tests and scripts.
+  const demo = m.contractAt("Demo", proxy);
 
-  // Upgrade the proxy to the new version of the Demo contract.
-  // This function also accepts a data parameter, which can be used to call a function,
-  // but we don't need it here so we pass an empty hex string ("0x").
-  m.call(proxyAdmin, "upgradeAndCall", [proxy, demoV2, "0x"], {
-    from: proxyAdminOwner,
-  });
-
-  // Return the proxy and proxy admin so that they can be used by other modules.
-  return { proxyAdmin, proxy };
+  // Return the contract instance, along with the original proxy and proxyAdmin contracts
+  // so that they can be used by other modules, or in tests and scripts.
+  return { demo, proxy, proxyAdmin };
 });
 
-/**
- * This is the third and final module that will be run.
- *
- * It takes the proxy from the previous module and uses it to create a local contract instance
- * for the DemoV2 contract. This allows us to interact with the DemoV2 contract via the proxy.
- */
-const interactableModule = buildModule("InteractableModule", (m) => {
-  // Get the proxy from the previous module.
-  const { proxy } = m.useModule(upgradeModule);
-
-  // Create a local contract instance for the DemoV2 contract.
-  // This line tells Hardhat Ignition to treat the contract at the proxy address as an DemoV2 contract.
-  // This allows us to call functions on the DemoV2 contract via the proxy.
-  const demo = m.contractAt("DemoV2", proxy);
-
-  // Return the contract instance so that it can be used by other modules or in tests.
-  return { demo };
-});
-
-module.exports = interactableModule;
+module.exports = demoModule;
