+++ examples/upgradeable/test/ProxyDemo.js
@@ -1,13 +1,24 @@
 const { expect } = require("chai");
 
 const ProxyModule = require("../ignition/modules/ProxyModule");
+const UpgradeModule = require("../ignition/modules/UpgradeModule");
 
 describe("Demo Proxy", function () {
+  describe("Proxy interaction", async function () {
+    it("Should be interactable via proxy", async function () {
+      const [owner, otherAccount] = await ethers.getSigners();
+
+      const { demo } = await ignition.deploy(ProxyModule);
+
+      expect(await demo.connect(otherAccount).version()).to.equal("1.0.0");
+    });
+  });
+
   describe("Upgrading", function () {
     it("Should have upgraded the proxy to DemoV2", async function () {
       const [owner, otherAccount] = await ethers.getSigners();
 
-      const { demo } = await ignition.deploy(ProxyModule);
+      const { demo } = await ignition.deploy(UpgradeModule);
 
       expect(await demo.connect(otherAccount).version()).to.equal("2.0.0");
     });
